navbar has a lot more functionality
moved adding movies to the navbar and removed the dedicated page (but kept the view)

moved searching movies to the navbar and removed the dedicated page (and removed the view).  results are displayed in the old all movie view, which has been renamed to movies to reflect its function better.